cleanup app generation in wallet, finish fullnode.sh staking
